fix: fixed display texture generating

+++ include/ui/widgets/Display.h
@@ -56,7 +56,7 @@ private:
         for (unsigned int i = 0; i < DISPLAY_WIDTH * DISPLAY_HEIGHT; i++) {
             pixels[i].position = {
                 static_cast<float>(i % DISPLAY_WIDTH),
-                static_cast<float>(i / DISPLAY_WIDTH)
+                static_cast<float>(static_cast<float>(i) / DISPLAY_WIDTH)
             };
             pixels[i].color = pixel_data.get_frame_buffer()[i] ?
                               static_cast<sf::Color>(true_color) : static_cast<sf::Color>(false_color);
